+++ docs/create_all_autodocs.py
@@ -1,11 +1,12 @@
 #!/usr/bin/env python
-# docs/create_all_autodocs.py
 
 """
-..
+docs/create_all_autodocs.py
 
 ===============================================================================
-    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
+
+    Copyright (C) 2012, University of Cambridge, Department of Psychiatry.
+    Created by Rudolf Cardinal (rnc1001@cam.ac.uk).
 
     This file is part of CamCOPS.
 
@@ -21,15 +22,14 @@
 
     You should have received a copy of the GNU General Public License
     along with CamCOPS. If not, see <https://www.gnu.org/licenses/>.
+
 ===============================================================================
 
-..
 """
 
 import argparse
 import logging
 import os
-import sys
 from typing import List
 
 from cardinal_pythonlib.fileops import rmtree
@@ -55,7 +55,8 @@ INDEX_FILENAME = "_index.rst"
 TOP_AUTODOC_INDEX = os.path.join(AUTODOC_DIR, INDEX_FILENAME)
 
 COPYRIGHT_COMMENT = r"""
-..  Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
+..  Copyright (C) 2012, University of Cambridge, Department of Psychiatry.
+    Created by Rudolf Cardinal (rnc1001@cam.ac.uk).
     .
     This file is part of CamCOPS.
     .