+++ server/camcops_server/alembic/versions/0061_fix_up_patient_idnum_ids.py
@@ -5,7 +5,8 @@ camcops_server/alembic/versions/0061_fix_up_patient_idnum_ids.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012, University of Cambridge, Department of Psychiatry.
+    Created by Rudolf Cardinal (rnc1001@cam.ac.uk).
 
     This file is part of CamCOPS.