+++ server/tools/build_server_translations.py
@@ -5,7 +5,8 @@ tools/build_server_translations.py
 
 ===============================================================================
 
-    Copyright (C) 2012-2020 Rudolf Cardinal (rudolf@pobox.com).
+    Copyright (C) 2012, University of Cambridge, Department of Psychiatry.
+    Created by Rudolf Cardinal (rnc1001@cam.ac.uk).
 
     This file is part of CamCOPS.
 
@@ -69,7 +70,7 @@ SOURCE_DIRS = [CAMCOPS_SERVER_DIRECTORY]  # it recurses below this directory
 
 PROJECT_NAME = "CamCOPS"
 COPYRIGHT_HOLDER = "Rudolf Cardinal"
-MSGID_BUGS_ADDR = "rudolf@pobox.com"
+MSGID_BUGS_ADDR = "rnc1001@cam.ac.uk"
 CHARSET = "utf-8"
 
 ENVVAR_POEDIT = "POEDIT"